requirements.txt
google-ads
streamlit==1.32.2
requests==2.31.0
python-dotenv==1.0.1
openai==1.14.3
selenium==4.20.0
webdriver-manager==4.0.1
